agentic-workflow/pyproject.toml
[project]
name = "agentic-workflow"
version = "0.1.0"
description = "Agentic Workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "nltk>=3.9.1",
    "python-dotenv>=1.0.1",
    "whisper>=1.1.10",
    "youtube-transcript-api>=0.6.3",
]
